+++ src/app/sitemap.ts
@@ -66,6 +66,7 @@ const pages = [
   { path: '/resources/public-safety-software-poland', priority: 0.75 },
   { path: '/resources/public-safety-software-belgium', priority: 0.75 },
   { path: '/resources/public-safety-software-norway', priority: 0.75 },
+  { path: '/resources/public-safety-software-denmark', priority: 0.75 },
   { path: '/resources/public-safety-software-uae', priority: 0.75 },
   { path: '/resources/public-safety-software-sweden', priority: 0.75 },
   { path: '/resources/public-safety-software-south-korea', priority: 0.75 },